requirements.txt
flask==1.0.2
flask-restful==0.3.7
gunicorn==19.9.0
mysql-connector-python==8.0.15
